requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
asttokens==3.0.0
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
backoff==2.2.1
billiard==4.2.1
blinker==1.9.0
boto3==1.39.10
botocore==1.39.10
cachetools==5.5.2
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
configobj==5.0.9
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
dacite==1.6.0
dagshub==0.6.2
dagshub-annotation-converter==0.1.11
databricks-sdk==0.60.0
dataclasses-json==0.6.7
debugpy==1.8.15
decorator==5.2.1
dictdiffer==0.9.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
dpath==2.2.0
dulwich==0.23.2
dvc==3.61.0
dvc-data==3.16.10
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
entrypoints==0.4
executing==2.2.0
fastapi==0.116.1
filelock==3.18.0
Flask==3.1.1
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.59.0
from-root==1.3.0
frozenlist==1.7.0
fsspec==2025.7.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
gql==3.5.3
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.3
gto==1.7.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hydra-core==1.3.2
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==7.2.1
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
kombu==5.5.4
lxml==6.0.0
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mlflow==2.15.0
mlflow-skinny==2.15.0
multidict==6.6.3
mypy-boto3-s3==1.39.5
mypy_extensions==1.1.0
narwhals==1.48.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==1.26.4
omegaconf==2.3.0
opentelemetry-api==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.11.0
packaging==24.2
pandas==2.3.1
parso==0.8.4
pathspec==0.12.1
pathvalidate==3.3.1
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pyarrow==15.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
pygit2==1.18.0
Pygments==2.19.2
pygtrie==2.5.0
pymongo==4.13.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.0
querystring-parser==1.2.4
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
s3transfer==0.13.1
scikit-learn==1.6.1
scipy==1.16.0
scmrepo==3.3.11
seaborn==0.13.2
semver==3.0.4
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
sqltrie==0.11.2
stack-data==0.6.3
starlette==0.47.2
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tomlkit==0.13.3
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
treelib==1.8.0
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
voluptuous==0.15.2
waitress==3.0.2
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.20.1
zc.lockfile==3.0.post1
zipp==3.23.0
